+++ src/Modules/Register/components/RegisterMailAndPassword.vue
@@ -9,25 +9,25 @@
       <sup>
         <i class="fas fa-asterisk" />
       </sup>
-    </div>
-    <div class="col-sm-auto">
       <input
         id="email"
         v-model="state.email"
         autocomplete="username"
-        :class="{ 'is-invalid': v$.email.$error }"
+        :class="{ 'is-invalid': v$.email.$error && v$.email.$dirty }"
         type="email"
         name="email"
         class="form-control"
+        :placeholder="$i18n('login.email_address')"
         @blur="update"
+        @input="v$.email.$touch()"
       >
       <div
-        v-if="v$.email.$error || !isMailValidForRegistration || isMailInvalid"
+        v-if="(v$.email.$error && v$.email.$dirty) || !isMailValidForRegistration || isMailInvalid"
         class="invalid-feedback"
       >
-        <span v-if="v$.email.required.$invalid">{{ $i18n('register.email_required') }}</span>
+        <span v-if="v$.email.required.$invalid && v$.email.$dirty">{{ $i18n('register.email_required') }}</span>
         <span v-else-if="v$.email.emailValidator.$invalid || !v$.email.foodsharing.$invalid || isMailInvalid">{{ $i18n('register.email_invalid') }}</span>
-        <span v-else-if="!isMailValidForRegistration">{{ $i18n('register.error_email_exist') }}</span>
+        <span v-else-if="!isMailValidForRegistration">{{ $i18n('register.error_email_invalid') }}</span>
       </div>
     </div>
     <div class="my-2">
@@ -38,58 +38,63 @@
             <i class="fas fa-asterisk" />
           </sup>
         </label>
-      </div>
-      <div class="col-sm-auto">
-        <input
+        <password-field
           id="password"
           v-model="state.password"
           autocomplete="new-password"
-          :class="{ 'is-invalid': v$.password.$error }"
+          :class="{ 'is-invalid': v$.password.$invalid && v$.password.$dirty }"
           type="password"
-          name="password"
-          class="form-control"
-          @input="emit('update:password', $event.target.value)"
-        >
-        <div v-if="v$.password.$error" class="invalid-feedback">
+          placeholder="login.password"
+          @input="v$.password.$touch()"
+        />
+        <div v-if="v$.password.$error && v$.password.$dirty" class="invalid-feedback">
           <span v-if="!v$.password.required">{{ $i18n('register.password_required') }}</span>
           <span v-if="!v$.password.minLength">{{ $i18n('register.password_minLength') }}</span>
         </div>
       </div>
-      <div class="my-1">
-        <div class="col-sm-auto">
-          <label for="confirmPassword">
-            {{ $i18n('register.login_passwd2') }}
-            <sup>
-              <i class="fas fa-asterisk" />
-            </sup>
-          </label>
-        </div>
-        <div class="col-sm-auto">
-          <input
-            id="confirmPassword"
-            v-model="state.confirmPassword"
-            autocomplete="new-password"
-            :class="{ 'is-invalid': v$.confirmPassword.$error }"
-            type="password"
-            name="confirmPassword"
-            class="form-control"
-          >
-          <div
-            v-if="v$.confirmPassword.$error"
-            class="invalid-feedback"
-          >
-            <span
-              v-if="!v$.confirmPassword.required"
-            >{{ $i18n('register.confirmPassword_required') }}</span>
-            <span
-              v-else-if="!v$.confirmPassword.sameAsPassword"
-            >{{ $i18n('register.confirmPassword_sameAsPassword') }}</span>
-          </div>
+      <div class="col-sm-auto">
+        <label for="confirmPassword">
+          {{ $i18n('register.login_passwd2') }}
+          <sup>
+            <i class="fas fa-asterisk" />
+          </sup>
+        </label>
+        <password-field
+          id="confirmPassword"
+          v-model="state.confirmPassword"
+          autocomplete="new-password"
+          :class="{ 'is-invalid': v$.confirmPassword.$invalid && v$.confirmPassword.$dirty }"
+          type="password"
+          placeholder="register.login_passwd2"
+          @input="v$.confirmPassword.$touch()"
+        />
+        <div
+          v-if="(v$.confirmPassword.$error && v$.confirmPassword.$dirty) || (v$.password.$error && v$.password.$dirty)"
+          class="invalid-feedback"
+        >
+          <ul>
+            <li v-if="v$.password.minLength.$invalid && v$.password.$dirty">
+              {{ $i18n('register.password_minLength') }}
+            </li>
+            <li v-if="v$.password.complexity.$invalid && v$.password.$dirty">
+              {{ $i18n('register.password_must_be_complex') }}
+            </li>
+            <li v-if="v$.password.isTrimmed.$invalid && v$.password.$dirty">
+              {{ $i18n('register.password_must_be_trimmed') }}
+            </li>
+            <li v-if="v$.confirmPassword.required.$invalid && v$.confirmPassword.$dirty">
+              {{ $i18n('register.confirmPassword_required') }}
+            </li>
+            <li v-if="v$.confirmPassword.sameAsPassword.$invalid && v$.confirmPassword.$dirty">
+              {{ $i18n('register.confirmPassword_sameAsPassword') }}
+            </li>
+          </ul>
         </div>
         <button
           class="btn btn-primary ml-3 mt-3"
           type="submit"
-          @click.prevent="redirect()"
+          :disabled="v$.$invalid || isMailInvalid || !isMailValidForRegistration"
+          @click.prevent="submit"
         >
           {{ $i18n('register.next') }}
         </button>
@@ -105,12 +110,13 @@
 </template>
 
 <script setup>
-import { ref, reactive, computed, defineEmits, defineProps } from 'vue'
+import { ref, reactive, computed, defineProps, defineEmits } from 'vue'
 import { useVuelidate } from '@vuelidate/core'
-import { required, email as emailValidator, minLength, sameAs, not } from '@vuelidate/validators'
+import { required, email as emailValidator, minLength, sameAs } from '@vuelidate/validators'
 import { testRegisterEmail } from '@/api/user'
-import { isFoodsharingDomain } from '@/helper/urls'
+import { isNotFoodsharingDomain } from '@/helper/urls'
 import { HTTP_RESPONSE } from '@/consts'
+import PasswordField from '@/components/Login/PasswordField.vue'
 
 const props = defineProps({
   email: { type: String, default: '' },
@@ -126,8 +132,8 @@ const state = reactive({
 })
 
 const validations = computed(() => ({
-  email: { required, emailValidator, foodsharing: not(sameAs(isFoodsharingDomain)) },
-  password: { required, minLength: minLength(8) },
+  email: { required, emailValidator, foodsharing: isNotFoodsharingDomain },
+  password: { required, minLength: minLength(8), isTrimmed: (value) => value.length === value.trim().length, complexity: (value) => /[a-z]/.test(value) && /[A-Z]/.test(value) && /[0-9]/.test(value) },
   confirmPassword: { required, sameAsPassword: sameAs(state.password) },
 }))
 
@@ -136,17 +142,17 @@ const v$ = useVuelidate(validations, state)
 const isMailValidForRegistration = ref(true)
 const isMailInvalid = ref(false)
 
-async function redirect () {
+async function submit () {
+  // Set password and email of the parent component
   await v$.value.$validate()
-  if (!v$.value.$invalid && !isMailInvalid.value && isMailValidForRegistration.value) {
-    emit('next')
-  }
+  emit('update:email', state.email)
+  emit('update:password', state.password)
+  emit('next')
 }
 
 async function update ($event) {
-  emit('update:email', $event.target.value)
   await v$.value.$validate()
-  isMailValidForRegistration.value = false
+  isMailValidForRegistration.value = true
   isMailInvalid.value = false
   try {
     const MailExist = await testRegisterEmail($event.target.value)
@@ -154,6 +160,7 @@ async function update ($event) {
   } catch (err) {
     if (err.code && err.code === HTTP_RESPONSE.BAD_REQUEST) {
       isMailInvalid.value = true
+      isMailValidForRegistration.value = false
       return isMailInvalid.value
     } else {
       throw err