+++ src/Modules/Settings/components/ChangeEmailForm.vue
@@ -9,13 +9,17 @@
         id="new-email"
         v-model="email"
         class="form-control mt-3"
-        :class="{ 'is-invalid': v$.email.$error }"
+        :class="{ 'is-invalid': v$.email.$invalid && v$.email.$dirty }"
         type="email"
         :placeholder="$i18n('settings.changemail.input_label_email')"
         :disabled="isLoading"
+        @input="v$.email.$touch()"
       >
-      <div v-if="v$.email.$error" class="invalid-feedback">
-        <span v-if="!v$.email.notFoodsharingAddress">
+      <div v-if="v$.email.$invalid && v$.email.$dirty" class="invalid-feedback">
+        <span v-if="v$.email.email">
+          {{ $i18n('settings.changemail.invalid') }}
+        </span>
+        <span v-else-if="v$.email.notFoodsharingAddress">
           {{ $i18n('settings.changemail.domain') }}
         </span>
       </div>
@@ -26,16 +30,17 @@
         id="new-email-confirm"
         v-model="confirmEmail"
         class="form-control mt-3"
-        :class="{ 'is-invalid': v$.confirmEmail.$error }"
+        :class="{ 'is-invalid': v$.confirmEmail.$invalid && v$.confirmEmail.$dirty }"
         type="email"
         :placeholder="$i18n('settings.changemail.input_label_email_confirm')"
         :disabled="isLoading"
+        @input="v$.confirmEmail.$touch()"
       >
       <div
-        v-if="v$.confirmEmail.$error"
+        v-if="v$.confirmEmail.$invalid && v$.confirmEmail.$dirty"
         class="invalid-feedback"
       >
-        <span v-if="!v$.confirmEmail.required || !v$.confirmEmail.sameAsEmail">
+        <span v-if="v$.confirmEmail.required || v$.confirmEmail.sameAsEmail">
           {{ $i18n('settings.changemail.confirm_email_required') }}
         </span>
       </div>
@@ -46,20 +51,26 @@
     </p>
 
     <div class="col-sm-auto">
-      <input
-        v-if="isMe"
-        id="password"
+      <password-field
         v-model="password"
-        class="form-control mt-3"
-        :class="{ 'is-invalid': v$.password.$error }"
+        class="mt-3"
+        :class="{ 'is-invalid': v$.password.$invalid && v$.password.$dirty }"
         type="password"
-        :placeholder="$i18n('settings.changemail.input_label_password')"
+        placeholder="settings.changemail.password_label"
         :disabled="isLoading"
+        @input="v$.password.$touch()"
+      />
+      <div
+        v-if="v$.password.$invalid && v$.password.$dirty"
+        class="invalid-feedback"
       >
+        {{ $i18n('settings.changemail.password_required') }}
+      </div>
     </div>
 
     <button
-      class="btn btn-primary btn-sm m-2 mt-3"
+      class="btn btn-sm m-2 mt-3"
+      :class="(isLoading || v$.$invalid) ? 'btn-secondary' : 'btn-primary'"
       :disabled="v$.$invalid"
       @click="submitEmail"
       v-text="$i18n('settings.email')"
@@ -70,15 +81,17 @@
 <script>
 import { pulseError, pulseInfo } from '@/script'
 import { useVuelidate } from '@vuelidate/core'
-import { email, minLength, not, required, requiredIf, sameAs } from '@vuelidate/validators'
+import { email, minLength, required, requiredIf, sameAs } from '@vuelidate/validators'
 import { requestEmailChange } from '@/api/settings'
-import { isFoodsharingDomain } from '@/helper/urls'
+import { isNotFoodsharingDomain } from '@/helper/urls'
 import { HTTP_RESPONSE } from '@/consts'
 import { useUserStore } from '@/stores/user'
+import PasswordField from '@/components/Login/PasswordField.vue'
 
 const userStore = useUserStore()
 
 export default {
+  components: { PasswordField },
   props: {
     /**
      * The form is visible for users themselves and for orga users. Users can only change their own email address by
@@ -104,15 +117,17 @@ export default {
       password: '',
     }
   },
-  validations: {
-    email: { required, minLength: minLength(1), email, notFoodsharingAddress: not(isFoodsharingDomain) },
-    confirmEmail: { required, minLength: minLength(1), email, sameAsEmail: sameAs('email') },
-    password: {
-      required: requiredIf(function () {
-        return this.isMe
-      }),
-      minLength: minLength(1),
-    },
+  validations () {
+    return {
+      email: { required, minLength: minLength(1), email, notFoodsharingAddress: isNotFoodsharingDomain },
+      confirmEmail: { required, minLength: minLength(1), email, sameAs: sameAs(this.email) },
+      password: {
+        required: requiredIf(function () {
+          return this.isMe
+        }),
+        minLength: minLength(1),
+      },
+    }
   },
   methods: {
     async submitEmail () {