+++ src/Modules/Settings/components/ChangePasswordForm.vue
@@ -8,13 +8,14 @@
       <password-field
         v-model="oldPassword"
         class="mt-3"
-        :class="{ 'is-invalid': v$.oldPassword.$error }"
+        :class="{ 'is-invalid': v$.oldPassword.$invalid && v$.oldPassword.$dirty }"
         type="password"
         placeholder="settings.change_password.old_password_label"
         :disabled="isLoading"
+        @input="v$.oldPassword.$touch()"
       />
       <div
-        v-if="v$.oldPassword.$error"
+        v-if="v$.oldPassword.$invalid && v$.oldPassword.$dirty"
         class="invalid-feedback"
       >
         {{ $i18n('settings.change_password.old_password_required') }}
@@ -25,21 +26,27 @@
       <password-field
         v-model="newPassword"
         class="mt-3"
-        :class="{ 'is-invalid': v$.newPassword.$error }"
+        :class="{ 'is-invalid': v$.newPassword.$invalid && v$.newPassword.$dirty }"
         type="password"
         placeholder="settings.change_password.new_password_label"
         :disabled="isLoading"
+        @input="v$.newPassword.$touch()"
       />
       <div
-        v-if="v$.newPassword.$error"
+        v-if="v$.newPassword.$invalid && v$.newPassword.$dirty"
         class="invalid-feedback"
       >
-        <span v-if="!v$.newPassword.required || !v$.newPassword.minLength">
-          {{ $i18n('settings.change_password.new_password_required') }}
-        </span>
-        <span v-else-if="!v$.newPassword.isTrimmed">
-          {{ $i18n('settings.change_password.new_password_must_be_trimmed') }}
-        </span>
+        <ul>
+          <li v-if="v$.newPassword.required.$invalid || v$.newPassword.minLength.$invalid">
+            {{ $i18n('settings.change_password.new_password_required') }}
+          </li>
+          <li v-if="v$.newPassword.complexity.$invalid">
+            {{ $i18n('settings.change_password.new_password_must_be_complex') }}
+          </li>
+          <li v-if="v$.newPassword.isTrimmed.$invalid">
+            {{ $i18n('settings.change_password.new_password_must_be_trimmed') }}
+          </li>
+        </ul>
       </div>
     </div>
 
@@ -47,13 +54,14 @@
       <password-field
         v-model="confirmNewPassword"
         class="mt-3"
-        :class="{ 'is-invalid': v$.confirmNewPassword.$error }"
+        :class="{ 'is-invalid': v$.confirmNewPassword.$invalid && v$.confirmNewPassword.$dirty }"
         type="password"
         placeholder="settings.change_password.new_password_confirmation_label"
         :disabled="isLoading"
+        @input="v$.confirmNewPassword.$touch()"
       />
       <div
-        v-if="v$.confirmNewPassword.$error"
+        v-if="v$.confirmNewPassword.$invalid && v$.confirmNewPassword.$dirty"
         class="invalid-feedback"
       >
         {{ $i18n('settings.change_password.new_password_confirmation_invalid') }}
@@ -61,7 +69,8 @@
     </div>
 
     <button
-      class="btn btn-primary btn-sm m-2 mt-3"
+      class="btn btn-sm m-2 mt-3"
+      :class="(isLoading || v$.$invalid) ? 'btn-secondary' : 'btn-primary'"
       :disabled="isLoading || v$.$invalid"
       @click="submitPassword"
       v-text="$i18n('settings.change_password.submit')"
@@ -91,10 +100,12 @@ export default {
       confirmNewPassword: '',
     }
   },
-  validations: {
-    oldPassword: { required, minLength: minLength(1) },
-    newPassword: { required, minLength: minLength(8), isTrimmed: (value) => value.length === value.trim().length },
-    confirmNewPassword: { required, sameAsPassword: sameAs('newPassword') },
+  validations () {
+    return {
+      oldPassword: { required, minLength: minLength(1) },
+      newPassword: { required, minLength: minLength(8), isTrimmed: (value) => value.length === value.trim().length, complexity: (value) => /[a-z]/.test(value) && /[A-Z]/.test(value) && /[0-9]/.test(value) },
+      confirmNewPassword: { required, sameAs: sameAs(this.newPassword) },
+    }
   },
   methods: {
     async submitPassword () {