+++ tests/Api/RegisterCept.php
@@ -14,7 +14,7 @@ $I->wantTo('Register a new user');
 $email = sq('email') . '@test.com';
 $first_name = sq('first_name');
 $last_name = sq('last_name');
-$pass = sq('pass');
+$pass = sq('pass') . 'abcABC123';
 $birthdate = '1990-05-31';
 
 $I->haveHttpHeader('Content-Type', 'application/x-www-form-urlencoded');